fixed some stuff and added const operator[]

+++ ex01/iter.hpp
@@ -2,16 +2,16 @@
 #define ITER_HPP
 
 template <typename T>
-void	iter(T *array, const int length, void (*func)(T &))
+void	iter(T *array, const size_t length, void (*func)(T &))
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 		func(array[i]);
 }
 
 template <typename T>
-void	iter(const T *array, const int length, void (*func)(const T &))
+void	iter(const T *array, const size_t length, void (*func)(const T &))
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 		func(array[i]);
 }
